set to double type for numbers

+++ api/src/db/models/fieldtypenumbervalue.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const FieldTypeNumberValue = sequelize.define('FieldTypeNumberValue', {
     orgId: DataTypes.INTEGER,
-    value: DataTypes.INTEGER
+    value: DataTypes.DOUBLE
   }, {})
   FieldTypeNumberValue.associate = function(models) {
     // associations can be defined here
